TST: splitting test_datetime.py
