+++ pandas/tests/indexes/datetime/test_datetime_misc.py
@@ -1,0 +1,48 @@
+import pandas.util.testing as tm
+from pandas import Float64Index, date_range, Timestamp
+
+
+class TestDateTimeIndexToJulianDate(tm.TestCase):
+
+    def test_1700(self):
+        r1 = Float64Index([2345897.5, 2345898.5, 2345899.5, 2345900.5,
+                           2345901.5])
+        r2 = date_range(start=Timestamp('1710-10-01'), periods=5,
+                        freq='D').to_julian_date()
+        self.assertIsInstance(r2, Float64Index)
+        tm.assert_index_equal(r1, r2)
+
+    def test_2000(self):
+        r1 = Float64Index([2451601.5, 2451602.5, 2451603.5, 2451604.5,
+                           2451605.5])
+        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
+                        freq='D').to_julian_date()
+        self.assertIsInstance(r2, Float64Index)
+        tm.assert_index_equal(r1, r2)
+
+    def test_hour(self):
+        r1 = Float64Index(
+            [2451601.5, 2451601.5416666666666666, 2451601.5833333333333333,
+             2451601.625, 2451601.6666666666666666])
+        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
+                        freq='H').to_julian_date()
+        self.assertIsInstance(r2, Float64Index)
+        tm.assert_index_equal(r1, r2)
+
+    def test_minute(self):
+        r1 = Float64Index(
+            [2451601.5, 2451601.5006944444444444, 2451601.5013888888888888,
+             2451601.5020833333333333, 2451601.5027777777777777])
+        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
+                        freq='T').to_julian_date()
+        self.assertIsInstance(r2, Float64Index)
+        tm.assert_index_equal(r1, r2)
+
+    def test_second(self):
+        r1 = Float64Index(
+            [2451601.5, 2451601.500011574074074, 2451601.5000231481481481,
+             2451601.5000347222222222, 2451601.5000462962962962])
+        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
+                        freq='S').to_julian_date()
+        self.assertIsInstance(r2, Float64Index)
+        tm.assert_index_equal(r1, r2)
